Improve seeds

+++ db/migrate/20161021000002_create_floorplans.rb
@@ -1,9 +1,9 @@
 class CreateFloorplans < ActiveRecord::Migration[5.0]
   def change
     create_table :floorplans do |t|
-      t.string   :name
-      t.text     :description
-      t.float    :rent
+      t.string  :name
+      t.text    :description
+      t.integer :rent
       t.integer :bathroom_count
       t.integer :bedroom_count
